requirements.txt
flask==3.0.0
scikit-learn==1.4.2
pandas==2.2.0
numpy==1.26.4
python-whois==0.9.4
